requirements.txt
discord
pysqlite3
python-dateutil